pyproject.toml
[tool.poetry]
name = "sshman"
version = "0.1.0"
description = ""
authors = ["Onigiri070 <4869909+onigiri070@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
